refactor: move file browser selection into the app database
