+++ src/server/src/services/app-service/app-service.hpp
@@ -82,6 +82,7 @@ public:
   std::shared_ptr<AbstractApplication> find(const QString &target) const;
   std::shared_ptr<AbstractApplication> findDefaultOpener(const QString &target) const;
   void setAdditionalSearchPaths(const std::vector<std::filesystem::path> &paths);
+  bool showInFileBrowser(const std::filesystem::path &path, bool select) const;
 
   bool openTarget(const QString &target) const;
   bool openTarget(const QUrl &target) const;
